controller/requirements.txt
requests
redis
joblib
scikit-learn
requests
numpy
torch
stable-baselines3
psutil
gymnasium
PyYAML
PyGithub
python-dotenv